+++ runtest.sh
@@ -135,9 +135,10 @@ rv="`$_cygwrun $srcdir/test/echoargs.bat /tmp /usr/a | tr -d '\r'`"
 test "x$rv" = "x$tmpdir $usrdir\\a" || xbexit 1 "Failed #7.1: \`$rv'"
 rv="`$_cygwrun $srcdir/test/echoargs.bat \"/tmp/some location\" | tr -d '\r'`"
 test "x$rv" = "x\"$tmpdir\\some location\"" || xbexit 1 "Failed #7.2: \`$rv'"
-rv="`$_cygwrun -t 2 $srcdir/test/infinite.bat`"
+$_cygwrun -i--- -t2 $srcdir/test/infinite.bat foo bar
 test $? -ne 2 && xbexit 1 "Failed #7.3"
-
+rv="`$_cygwrun $srcdir/test/echoargs.bat /tmp /usr/b | tr -d '\r'`"
+test "x$rv" = "x$tmpdir $usrdir\\b" || xbexit 1 "Failed #7.4: \`$rv'"
 
 echo "All tests passed!"
 exit 0